# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GESTÃO DE BIBLIOTECA",
    "description": "Aplicativo de gerenciamento de empréstimo de livros em uma biblioteca.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/users": {
      "get": {
        "summary": "Rota de listagem de usuários",
        "description": "Apenas funcionários podem acessa esta rota",
        "tags": ["Users"],
        "operationId": "listUsers",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Listagem de usuários com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Usuário não está autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Usuário não tem permissão para está ação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "summary": "Detalhes do usuário",
        "description": "Apenas funcionários e o dono da conta podem acessa esta rota",
        "tags": ["Users"],
        "operationId": "retrieveUser",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "ID do usuário",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Usuário não está autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Usuário não tem permissão para está ação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Atualização de usuário",
        "description": "Apenas funcionários e o dono da conta podem acessa esta rota",
        "tags": ["Users"],
        "operationId": "updateUser",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "ID do usuário",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": {
            "description": "Usuário não está autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Usuário não tem permissão para está ação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Deleção de usuário",
        "description": "Apenas funcionários e o dono da conta podem acessa esta rota",
        "tags": ["Users"],
        "operationId": "deleteUser",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "ID do usuário",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Usuário deletado com sucesso"
          },
          "401": {
            "description": "Usuário não está autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Usuário não tem permissão para está ação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "is_staff": {
            "type": "boolean"
          },
          "is_blocked_loans": {
            "type": "boolean",
            "default": false
          },
          "date_unlock": {
            "type": "string",
            "default": null
          }
        }
      },
      "UpdateUserPayload": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "email": { "type": "string" },
          "password": { "type": "string" }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "apiKey": "biblioteca_api_test",
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}